+++ src/main/java/com/friendlyblood/api/utils/exceptions/InternalServerException.java
@@ -1,0 +1,11 @@
+package com.friendlyblood.api.utils.exceptions;
+
+public class InternalServerException extends RuntimeException {
+    public InternalServerException(String message) {
+        super(message);
+    }
+
+    public InternalServerException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
